phase3/chatkit/backend/main.py
"""ChatKit Backend - Entry point."""

if __name__ == "__main__":
    import uvicorn
    import os
    from dotenv import load_dotenv

    load_dotenv()

    HOST = os.getenv("HOST", "0.0.0.0")
    PORT = int(os.getenv("PORT", "8002"))

    print(f"[INFO] Starting ChatKit Backend on {HOST}:{PORT}")

    uvicorn.run(
        "src.main:app",
        host=HOST,
        port=PORT,
        reload=True,
        log_level="info"
    )
